feat: add error handling and exception

+++ app/Http/Controllers/Api/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class PostController extends Controller
 {
@@ -17,25 +18,52 @@ class PostController extends Controller
     public function show($id)
     {
         $data = Post::find($id);
+
+        if (is_null($data)) {
+            return response()->json([
+                'message' => 'Resource not found!'
+            ], 404);
+        }
+
         return response()->json($data, 200);
     }
 
     public function store(Request $request)
     {
-        $data = Post::create($request->all());
+        $validatedData = $request->validate([
+            'title' => 'required|string|min:5',
+        ]);
+
+        $data = Post::create($validatedData);
+
         return response()->json($data, 201);
     }
+
 
     public function update(Request $request, $id)
     {
         $data = Post::find($id);
         $data->update($request->all());
+
+        if (is_null($data)) {
+            return response()->json([
+                'message' => 'Resource not found!'
+            ], 404);
+        }
+
         return response()->json($data, 200);
     }
 
     public function destroy($id)
     {
         $data = Post::find($id);
+
+        if (is_null($data)) {
+            return response()->json([
+                'message' => 'Resource not found!'
+            ], 404);
+        }
+
         $data->delete();
         return response()->json(null, 204);
     }
